requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
aioredis==1.3.1
python-dotenv
numpy==1.18.5
nltk==3.5
emoji
